Add: MainPageView
