+++ .gitignore
@@ -403,4 +403,4 @@ zsdoc/data
 my_settings.py
 
 db_uploader.py
-products.csv
+csv
