+++ guroom/urls.py
@@ -1,6 +1,5 @@
 from django.urls import path, include
 
 urlpatterns = [
-    path('', include('main.urls')),
     path('products', include('products.urls'))
 ]
