--- main/urls.py
@@ -1,7 +1,0 @@
-from django.urls import path
-
-from main.views import MainPageView
-
-urlpatterns = {
-    path('', MainPageView.as_view()),
-}
